started the transactions operations

+++ src/main/java/com/dgsaltarin/backend/domain/dto/CreateAccountRequest.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class CreateAccountRequest {
 
     private String userName;
-    private double initialBalance;
+    private int initialBalance;
 }
