+++ src/main/java/com/dgsaltarin/backend/domain/dto/CreatedAccountResponse.java
@@ -3,6 +3,6 @@ package com.dgsaltarin.backend.domain.dto;
 public class CreatedAccountResponse {
 
     private String userName;
-    private double initialBalance;
+    private int initialBalance;
     private int accountNumber;
 }
